// superset-frontend/src/components/AlteredSliceTag/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import type { QueryFormData } from '@superset-ui/core';

export interface AlteredSliceTagProps {
  origFormData: QueryFormData;
  currentFormData: QueryFormData;
}

export interface ControlMap {
  [key: string]: {
    label?: string;
    type?: string;
  };
}

export type FilterItemType = {
  comparator?: string | string[];
  subject: string;
  operator: string;
  label?: string;
};

export type DiffItemType<
  T = FilterItemType | number | string | Record<string | number, any>,
> =
  | T[]
  | boolean
  | number
  | string
  | Record<string | number, any>
  | null
  | undefined;

export type DiffType = {
  before: DiffItemType;
  after: DiffItemType;
};

export type RowType = {
  before: string | number;
  after: string | number;
  control: string;
};